yujun_test/requirements.txt
streamlit==1.28.1
openai==1.3.7
python-dotenv==1.0.0 